Fixed a number of implicit conversion warnings.

+++ Source/MIKMIDIControlChangeCommand.m
@@ -83,7 +83,7 @@
 - (void)setControllerNumber:(NSUInteger)value
 {
 	if (![[self class] isMutable]) return MIKMIDI_RAISE_MUTATION_ATTEMPT_EXCEPTION;
-	self.dataByte1 = value;
+	self.dataByte1 = (UInt8)value;
 }
 
 - (NSUInteger)controllerValue { return self.value; }
