+++ Source/MIKMIDIProgramChangeCommand.m
@@ -41,7 +41,7 @@
 - (void)setProgramNumber:(NSUInteger)value
 {
 	if (![[self class] isMutable]) return MIKMIDI_RAISE_MUTATION_ATTEMPT_EXCEPTION;
-	self.dataByte1 = value;
+	self.dataByte1 = (UInt8)value;
 }
 
 @dynamic channel; // MIKMIDIChannelVoiceCommand already implements a getter *and* setter for this. This keeps the compiler happy.
